backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
# psycopg2-binary==2.9.9  # PostgreSQL driver - uncomment if using PostgreSQL
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
msal==1.26.0
PyPDF2==3.0.1
python-docx==1.1.0
textract==1.6.5
openai==1.3.8
langchain==0.0.350
langchain-openai==0.0.2
boto3==1.34.0
cryptography==41.0.8
bleach==6.1.0
celery==5.3.4
redis==5.0.1